+++ docs/changelog.md
@@ -7,6 +7,7 @@
 + Fixed small error in CL650 example mapping
 + Added default encoder mode in device settings
 + Moved logfile into preference folder where it should have been from the beginning on
++ Added reload button on general page of aircraft profile
 
 -----------------------------------------------------------------------------------------------------------------------